Include full list of supposed GitHub reasons
I left out a few reasons originally because I thought perhaps they were
only valid for the Notifications API, not emails. Turns out 'manual' is
indeed a reason mailed out, so this allows that reason in Dug.

I added the other ones as well, but I'm not convinced 'subscribed' is a
thing in emails.