Manually reverse message order from Gmail API
This mitigates trouble processing state changes because of message
order.  The Gmail API always provides them date descending and gives no
order querying (indeed, with labels they don't even allow you to sort in
the Gmail UI!). So if someone closes/reopens an issue, the reopen will
be processed first and closed after, resulting in a final state
of...closed.

This could possibly be corrected in the future by taking the entire
thread in to account if it's a state change, but with the
message-by-message implementation currently set up, this is a hack that
will fix *most* cases.